# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bff_message",
    "version": "1.0.0"
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "PAYLOAD_TOO_LARGE",
                  "UNPROCESSABLE_ENTITY",
                  "TOO_MANY_REQUESTS",
                  "INTERNAL_ERROR",
                  "BAD_GATEWAY",
                  "SERVICE_UNAVAILABLE",
                  "GATEWAY_TIMEOUT"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "path": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/attachments": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Télécharge une pièce jointe",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "description": "Fichier à télécharger en tant que pièce jointe",
                    "example": "fichier.pdf"
                  }
                },
                "description": "Corps de la requête pour télécharger une pièce jointe"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pièce jointe téléchargée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "attachments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "number"
                              }
                            ],
                            "description": "Identifiant unique, peut être une chaîne ou un nombre",
                            "example": "12345"
                          },
                          "name": {
                            "type": "string",
                            "description": "Nom de la pièce jointe",
                            "example": "document.pdf"
                          },
                          "size": {
                            "type": "integer",
                            "description": "Taille de la pièce jointe en octets",
                            "example": 102400
                          },
                          "type": {
                            "type": "string",
                            "description": "Type MIME de la pièce jointe",
                            "example": "application/pdf"
                          },
                          "url": {
                            "type": "string",
                            "format": "uri",
                            "description": "URL pour accéder à la pièce jointe",
                            "example": "https://example.com/document.pdf"
                          }
                        },
                        "required": [
                          "id",
                          "name"
                        ],
                        "description": "Représentation d’une pièce jointe"
                      },
                      "description": "Liste des pièces jointes téléchargées"
                    }
                  },
                  "required": [
                    "attachments"
                  ],
                  "description": "Réponse contenant la liste des pièces jointes téléchargées"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Core API is unavailable or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}